# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation TEA BLISS",
        "description": "API với Sanctum Authentication của Tea Bliss",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Lấy danh sách bài viết",
                "operationId": "63a0653b902e8cc3239e3b0fa8d89b07",
                "responses": {
                    "200": {
                        "description": "Lấy danh sách bài viết thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy danh sách bài viết thành công"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Bài viết về công nghệ"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "bai-viet-ve-cong-nghe"
                                                    },
                                                    "thumbnail_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/image.jpg"
                                                    },
                                                    "short_desciption": {
                                                        "type": "string",
                                                        "example": "Tóm tắt bài viết"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "Nội dung chi tiết bài viết"
                                                    },
                                                    "product_id": {
                                                        "type": "integer",
                                                        "example": 5,
                                                        "nullable": true
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "article_category_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "status": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-03-06T12:00:00Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-03-06T12:30:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lấy danh sách bài viết thất bại",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy danh sách bài viết thất bại"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Lỗi server"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Lấy chi tiết bài viết",
                "operationId": "0ea1b47d994cf0b1a6e86dc71f1146a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của bài viết",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lấy chi tiết bài viết thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy chi tiết bài viết thành công"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Bài viết về công nghệ"
                                                },
                                                "slug": {
                                                    "type": "string",
                                                    "example": "bai-viet-ve-cong-nghe"
                                                },
                                                "thumbnail_url": {
                                                    "type": "string",
                                                    "example": "https://example.com/image.jpg"
                                                },
                                                "short_desciption": {
                                                    "type": "string",
                                                    "example": "Tóm tắt bài viết"
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "Nội dung chi tiết bài viết"
                                                },
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 5,
                                                    "nullable": true
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "article_category_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "status": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-06T12:00:00Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-03-06T12:30:00Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tìm thấy bài viết",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy chi tiết bài viết thất bại"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Không tìm thấy"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy chi tiết bài viết thất bại"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Lỗi server"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}